+++ server/validations/appVersion.validation.js
@@ -27,33 +27,6 @@ let getLatestVersion = (req, res, next) => {
   else next();
 };
 
-let updateMessage = (req, res, next) => {
-  let updateMessageSchema = joi.object().keys({
-    message: joi.string().required(),
-    updatedMessage: joi.string().required(),
-    updatedOn: joi.date()
-  });
-  let { error } = joi.validate(req.body, updateMessageSchema);
-  if (error) res.status(422).send({
-    success: false,
-    error
-  });
-  else next();
-};
-
-let addMessage = (req, res, next) => {
-  let addMessageSchema = joi.object().keys({
-    message: joi.string(),
-    updatedOn: joi.date()
-  });
-  let { error } = joi.validate(req.body, addMessageSchema);
-  if (error) res.status(422).send({
-    success: false,
-    error
-  });
-  else next();
-};
-
 let listAllVersions = (req, res, next) => {
   let updatedVersionsSchema = joi.object().keys({
     appCode: joi.string().required()
@@ -84,6 +57,4 @@ module.exports = {
   getUpdatedVersions,
   getLatestVersion,
   listAllVersions,
-  addMessage,
-  updateMessage
 };
